Refactored receiver and sender, normal vs ondemand, TCP vs MPI
