+++ ff/distributed/ff_dsender.hpp
@@ -23,20 +23,27 @@
 using namespace ff;
 
 class ff_dsender: public ff_minode_t<message_t> { 
-private:
+protected:
     size_t neos=0;
     int next_rr_destination = 0; //next destiation to send for round robin policy
     std::vector<ff_endpoint> dest_endpoints;
     std::map<int, int> dest2Socket;
-    std::map<int, int> sockets;
+    std::vector<int> sockets;
 	int coreid;
 
     int receiveReachableDestinations(int sck){
+       
         size_t sz;
+        ssize_t r;
 
-        //while (readvn(sck, iov, 1) != sizeof(sz)) // wait untill a size is received!
-        recv(sck, &sz, sizeof(sz), MSG_WAITALL);
-
+        if ((r=readn(sck, (char*)&sz, sizeof(sz)))!=sizeof(sz)) {
+            if (r==0)
+                error("Error unexpected connection closed by receiver\n");
+            else			
+                error("Error reading size (errno=%d)");
+            return -1;
+        }
+	
         sz = be64toh(sz);
 
         std::cout << "Receiving routing table (" << sz << " bytes)" << std::endl;
@@ -114,14 +121,15 @@ private:
         #endif
 
         // receive the reachable destination from this sockets
-        receiveReachableDestinations(socketFD);
+        if (receiveReachableDestinations(socketFD) < 0)
+            return -1;
 
 
         return socketFD;
     }
 
     int tryConnect(const ff_endpoint &destination){
-        int fd, retries = 0;
+        int fd = -1, retries = 0;
 
         while((fd = this->create_connect(destination)) < 0 && ++retries < MAX_RETRIES)
             if (retries < AGGRESSIVE_TRESHOLD)
@@ -134,7 +142,6 @@ private:
     }
 
     int sendToSck(int sck, message_t* task){
-        //std::cout << "received something from " << task->sender << " directed to " << task->chid << std::endl;
         task->sender = htonl(task->sender);
         task->chid = htonl(task->chid);
 
@@ -171,6 +178,7 @@ public:
 		if (coreid!=-1)
 			ff_mapThreadToCpu(coreid);
 		
+        sockets.resize(this->dest_endpoints.size());
         for(size_t i=0; i < this->dest_endpoints.size(); i++)
             if ((sockets[i] = tryConnect(this->dest_endpoints[i])) <= 0 ) return -1;
 
@@ -180,8 +188,9 @@ public:
     void svc_end() {
         // close the socket not matter if local or remote
         for(size_t i=0; i < this->sockets.size(); i++)
-            close(sockets[i]);    
+            close(sockets[i]);   
     }
+
     message_t *svc(message_t* task) {
         /* here i should send the task via socket */
         if (task->chid == -1){ // roundrobin over the destinations
@@ -190,7 +199,6 @@ public:
         }
 
         sendToSck(dest2Socket[task->chid], task);
-
         delete task;
         return this->GO_ON;
     }
@@ -200,11 +208,161 @@ public:
             message_t * E_O_S = new message_t;
             E_O_S->chid = 0;
             E_O_S->sender = 0;
-            for(const auto& pair : sockets)
-                sendToSck(pair.second, E_O_S);
+            for(const auto& sck : sockets)
+                sendToSck(sck, E_O_S);
 
             delete E_O_S;
         }
+    }
+
+};
+
+
+
+/*
+    ONDEMAND specification
+*/
+
+class ff_dsenderOD: public ff_dsender { 
+private:
+    int last_rr_socket = 0; //next destiation to send for round robin policy
+    std::map<int, unsigned int> sockCounters;
+    const int queueDim;
+    fd_set set, tmpset;
+    int fdmax = -1;
+
+    
+public:
+    ff_dsenderOD(ff_endpoint dest_endpoint, int queueDim = 1, int coreid=-1)
+		: ff_dsender(dest_endpoint, coreid), queueDim(queueDim) {}
+
+    ff_dsenderOD(std::vector<ff_endpoint> dest_endpoints_, int queueDim = 1, int coreid=-1)
+		: ff_dsender(dest_endpoints_, coreid), queueDim(queueDim) {}
+
+    int svc_init() {
+		if (coreid!=-1)
+			ff_mapThreadToCpu(coreid);
+		
+        FD_ZERO(&set);
+        FD_ZERO(&tmpset);
+
+		sockets.resize(this->dest_endpoints.size());
+        for(size_t i=0; i < this->dest_endpoints.size(); i++){
+            if ((sockets[i] = tryConnect(this->dest_endpoints[i])) <= 0 ) return -1;
+            
+            // execute the following block only if the scheduling is onDemand
+            
+            sockCounters[sockets[i]] = queueDim;
+            FD_SET(sockets[i], &set);
+            if (sockets[i] > fdmax)
+                fdmax = sockets[i];
+            
+        }
+
+        return 0;
+    }
+
+    int waitAckFrom(int sck){
+        while (sockCounters[sck] == 0){
+            for (size_t i = 0; i < this->sockets.size(); ++i){
+                int r; ack_t a;
+                if ((r = recvnnb(sockets[i], reinterpret_cast<char*>(&a), sizeof(ack_t))) != sizeof(ack_t)){
+                    if (errno == EWOULDBLOCK){
+                        assert(r == -1);
+                        continue;
+                    }
+                    perror("recvnnb ack");
+                    return -1;
+                } else 
+                    //printf("received ACK from conn %d\n", i);
+                    sockCounters[sockets[i]]++;
+                
+            }
+			
+            if (sockCounters[sck] == 0){
+                tmpset = set;
+                if (select(fdmax + 1, &tmpset, NULL, NULL, NULL) == -1){
+                    perror("select");
+                    return -1;
+                }
+            }
+        }
+        return 1;
+    }
+
+	int waitAckFromAny() {
+		tmpset = set;
+		if (select(fdmax + 1, &tmpset, NULL, NULL, NULL) == -1){
+			perror("select");
+			return -1;
+		} 
+		// try to receive from all connections in a non blocking way
+		for (size_t i = 0; i < this->sockets.size(); ++i){
+			int r; ack_t a;
+			int sck = sockets[i];
+			if ((r = recvnnb(sck, reinterpret_cast<char*>(&a), sizeof(ack_t))) != sizeof(ack_t)){
+				if (errno == EWOULDBLOCK){
+					assert(r == -1);
+					continue;
+				}
+				perror("recvnnb ack");
+				return -1;
+			} else {
+				sockCounters[sck]++;
+				last_rr_socket = i;
+				return sck;
+			}
+		} 
+		assert(1==0);
+		return -1;
+	}
+
+    int getNextReady(){
+        for(size_t i = 0; i < this->sockets.size(); i++){
+            int actualSocket = (last_rr_socket + 1 + i) % this->sockets.size();
+            int sck = sockets[actualSocket];
+            if (sockCounters[sck] > 0) {
+                last_rr_socket = actualSocket;
+                return sck;
+            }
+        }
+		return waitAckFromAny();		
+    }
+
+
+    void svc_end() {
+		long totalack = sockets.size()*queueDim;
+		long currack  = 0;
+        for(const auto& pair : sockCounters)
+            currack += pair.second;
+		while(currack<totalack) {
+			waitAckFromAny();
+			currack++;
+		}
+
+		// close the socket not matter if local or remote
+        for(size_t i=0; i < this->sockets.size(); i++) {
+	            close(sockets[i]);
+		}
+    }
+    message_t *svc(message_t* task) {
+        int sck;
+        if (task->chid != -1){
+            sck = dest2Socket[task->chid];
+            if (sockCounters[sck] == 0 && waitAckFrom(sck) == -1){ // blocking call if scheduling is ondemand
+                    error("Error waiting Ack from....\n");
+                    delete task; return this->GO_ON;
+            }
+        } else 
+            sck = getNextReady(); // blocking call if scheduling is ondemand
+    
+        sendToSck(sck, task);
+
+        // update the counters
+        sockCounters[sck]--;
+
+        delete task;
+        return this->GO_ON;
     }
